I think I got it working!

+++ pkg/types/fakes/valid_measurement_conversion.go
@@ -14,7 +14,7 @@ func BuildFakeValidMeasurementConversion() *types.ValidMeasurementConversion {
 		From:              *BuildFakeValidMeasurementUnit(),
 		To:                *BuildFakeValidMeasurementUnit(),
 		OnlyForIngredient: nil,
-		Modifier:          BuildFakeNumber(),
+		Modifier:          float32(BuildFakeNumber()),
 		Notes:             buildUniqueString(),
 		CreatedAt:         fake.Date(),
 	}
